chore: changed interactiontext to resourceheading
