--- apps/watch/src/components/InteractionText/index.ts
@@ -1,1 +1,0 @@
-export { InteractionText } from './InteractionText'
