+++ apps/watch/src/components/ResourceHeading/index.ts
@@ -1,0 +1,1 @@
+export { ResourceHeading } from './ResourceHeading'
