+++ apps/watch/src/components/StrategiesView/StrategiesView.tsx
@@ -3,7 +3,7 @@ import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
 import Stack from '@mui/material/Stack'
 import { ReactElement } from 'react'
-import { InteractionText } from '../InteractionText'
+import { ResourceHeading } from '../ResourceHeading'
 import { StrategySections } from '../StrategySections'
 
 export function StrategiesView(): ReactElement {
@@ -11,7 +11,7 @@ export function StrategiesView(): ReactElement {
     <Box sx={{ backgroundColor: 'background.default' }}>
       <Container maxWidth="xxl">
         <Stack sx={{ p: 4, gap: 4 }}>
-          <InteractionText heading="Resource" />
+          <ResourceHeading heading="Resource" />
           <SearchBar />
           <StrategySections />
         </Stack>
